backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
vaderSentiment==3.3.2
textstat==0.7.4
